requirements.txt
prettytable==3.11.0
psycopg2-binary==2.9.9
wcwidth==0.2.13
